readme changes

+++ dataset_interface.py
@@ -49,7 +49,7 @@ def custom_collate(data):
 class Data_Tuple():
     #basically a struct
     def __init__(self, imgL : torch.Tensor, imgR : torch.Tensor, depth_gtL : torch.Tensor, depth_gtR : torch.Tensor, focalLength : torch.Tensor, baseline : torch.Tensor) -> None:
-        self.imgL = torch.FloatTensor(imgL)
+        self.imgL = imgL
         self.imgR = imgR
         self.depthL = depth_gtL
         self.depthR = depth_gtR
